modules/location-api/requirements.txt
grpcio
kafka-python
marshmallow
protobuf
